container_manager_app/requirements.txt
docker
mysql-connector-python